requirements.txt
# opencv-python
Django
Pillow